create booking partially done

+++ src/main/java/com/pramu_k/easybook/repository/FlightRepository.java
@@ -3,5 +3,5 @@ package com.pramu_k.easybook.repository;
 import com.pramu_k.easybook.model.Flight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlightRepository extends JpaRepository<Flight, Integer> {
+public interface FlightRepository extends JpaRepository<Flight, Long> {
 }
